MNIST-server/requirements.txt
astroid==2.0.4
cloudpickle==0.6.1
colorama==0.4.0
dask==0.20.0
decorator==4.3.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
networkx==2.2
numpy==1.15.3
opencv-python==3.4.3.18
Pillow==5.3.0
pylint==2.1.1
PyWavelets==1.0.1
pyzmq==17.1.2
scikit-image==0.14.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
toolz==0.9.0
typed-ast==1.1.0
wrapt==1.10.11
zmq==0.0.0
